requirements.txt
# Utility packages
aiohttp>=3.10.10
pytchat==0.5.5
requests>=2.32.3  # For API testing
python-dotenv==1.0.1
tqdm>=4.66.1
fastapi>=0.115.6
uvicorn[standard]>=0.25.0
pyjwt>=2.10.1  # For stream auth
pyttsx3==2.90
pydub==0.25.1
tenacity>=8.5.0
textblob==0.17.1
pytest==7.4.0
pytest-asyncio==0.21.1
sounddevice>=0.4.6
numpy>=1.24.0
wave>=0.0.2
rich>=13.0.0
web3==6.4.0 # for future crypto integrations
eth-account==0.13.4
eth-typing==5.0.1
eth-utils==5.1.0
async-timeout==5.0.1
cytoolz>=0.12.2
eth-hash[pycryptodome]>=0.5.2
keyboard>=0.13.5

# VTube Studio Integration
pyvts==0.3.3
aiofiles>=0.6.0
websockets>=10.0

# Database drivers
neo4j>=5.14.0
motor>=3.3.2
pytz>=2024.2 # for timezone support
pydantic>=2.10.6

# AI/ML packages
openai>=1.12.0
google-generativeai>=0.8.3
together>=0.2.0
anthropic>=0.37.1
voyageai>=0.3.2
groq>=0.13.1
elevenlabs>=0.3.0

# LangChain and related packages
langchain-google-genai>=2.0.1
langchain>=0.3.4
langchain-core>=0.3.12
langchain-experimental>=0.0.49
langchain-community>=0.0.13
langchain-openai>=0.2.3
langchain-anthropic>=0.2.3

# Message processing
tiktoken>=0.5.2